pyproject.toml
[project]
name = "wind"
version = "0.1.0"
description = "Prediction of wind power using weather forecast"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cf-xarray>=0.10.7",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.5",
    "notebook>=7.4.5",
    "optuna>=4.5.0",
    "optuna-dashboard>=0.19.0",
    "optuna-integration[xgboost]>=4.5.0",
    "plotly>=6.2.0",
    "polars>=1.32.1",
    "pyarrow>=21.0.0",
    "rioxarray>=0.19.0",
    "ruff>=0.12.7",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "statsforecast>=2.0.1",
    "statsmodels>=0.14.5",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
    "xarray>=2025.7.1",
    "xgboost>=3.0.4",
    "zarr>=3.1.1",
]

[tool.ruff]
extend-include = ["*.ipynb"]

[dependency-groups]
bayes = [
    "pymc>=5.25.1",
]
geopandas = [
    "geopandas>=1.1.1",
    "folium>=0.20.0",
    "mapclassify>=2.10.0",
    "cartopy>=0.25.0",
]

[project.scripts]
wind = "wind:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"